# IT_DBMS_Python_JS/src/openapi_server/models/edit_value_location.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class EditValueLocation(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    EditValueLocation - a model defined in OpenAPI

        row_id: The row_id of this EditValueLocation.
        column_name: The column_name of this EditValueLocation.
    """

    row_id: int = Field()
    column_name: str = Field()
    value: Any = Field()

    @validator("row_id")
    def row_id_min(cls, value):
        assert value >= 1
        return value

    @validator("column_name")
    def column_name_min_length(cls, value):
        assert len(value) >= 1
        return value

EditValueLocation.update_forward_refs()